+++ .github/workflows/ci.yml
@@ -43,10 +43,6 @@ jobs:
       - name: Build Lib
         run: |
           pnpm build
-      - name: Build Docs plugin
-        run: |
-          cd packages/vitepress/vitepress-demo-block
-          pnpm run build
       - name: Build Docs
         run: |
           cd packages/hooks
